Automatizacion/Reemplazar_formato/requirements.txt
Flask==3.1.0
pandas==2.2.3
python-docx==1.1.2
openpyxl==3.1.5
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.6
docxtpl==0.19.1
docxcompose==1.4.0
lxml==5.3.1
Unidecode==1.3.8
